+++ src/Libraries/CoreNodes/DSCoreNodes.Migrations.xml
@@ -26,27 +26,27 @@
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.CSV.WriteToFile</oldName>
-    <newName>DSCore.Data.ExportCSV</newName>
+    <newName>ImportExport.Data.ExportCSV</newName>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.Directory.Contents</oldName>
-    <newName>DSCore.IO.File.GetDirectoryContents</newName>
+    <newName>ImportExport.File.GetDirectoryContents</newName>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.Directory.Copy</oldName>
-    <newName>DSCore.IO.File.CopyDirectory</newName>
+    <newName>ImportExport.File.CopyDirectory</newName>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.Directory.Delete</oldName>
-    <newName>DSCore.IO.File.DeleteDirectory</newName>
+    <newName>ImportExport.File.DeleteDirectory</newName>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.Directory.Exists</oldName>
-    <newName>DSCore.IO.File.DirectoryExists</newName>
+    <newName>ImportExport.File.DirectoryExists</newName>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.IO.Directory.Move</oldName>
-    <newName>DSCore.IO.File.MoveDirectory</newName>
+    <newName>ImportExport.File.MoveDirectory</newName>
   </priorNameHint>
   <priorNameHint>
     <oldName>DSCore.Display.ByGeometryColor</oldName>
